+++ go2shop-catalogue-service/src/main/java/com/go2shop/catalogue/entity/Product.java
@@ -11,48 +11,47 @@ import javax.persistence.Table;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
-/** 
+/**
  * 
- * <Write a short description on the purpose of the class> 
+ * <Write a short description on the purpose of the class>
  * 
- * @author P1326154 
- * Created Date Aug 10, 2021 3:20:15 PM 
+ * @author P1326154 Created Date Aug 10, 2021 3:20:15 PM
  * 
-*/
+ */
 @Entity
 @Table(name = "TB_PRODUCT")
 public class Product {
 
 	@Id
-	@GeneratedValue(strategy=GenerationType.IDENTITY)
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "ID")
-	private int id;
-	
+	private Long id;
+
 	@NotBlank
 	@Column(name = "Name")
 	private String name;
-	
+
 	@NotBlank
 	@Column(name = "Description")
 	private String description;
-	
+
 	@NotNull
 	@Column(name = "Price")
 	private BigDecimal price;
-	
+
 	@NotNull
 	@Column(name = "Stock")
 	private int stock;
-	
+
 	@NotNull
 	@Column(name = "TB_USER_ID")
-	private int userID;
+	private Long userID;
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -88,12 +87,12 @@ public class Product {
 		this.stock = stock;
 	}
 
-	public int getUserID() {
+	public Long getUserID() {
 		return userID;
 	}
 
-	public void setUserID(int userID) {
+	public void setUserID(Long userID) {
 		this.userID = userID;
 	}
-	
+
 }
