+++ go2shop-order-service/src/main/java/com/go2shop/order/entity/Order.java
@@ -11,40 +11,39 @@ import javax.persistence.Table;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
-/** 
+/**
  * 
- * <Write a short description on the purpose of the class> 
+ * <Write a short description on the purpose of the class>
  * 
- * @author P1326154 
- * Created Date Aug 10, 2021 3:06:42 PM 
+ * @author P1326154 Created Date Aug 10, 2021 3:06:42 PM
  * 
-*/
+ */
 @Entity
 @Table(name = "TB_ORDER")
 public class Order {
 
 	@Id
-	@GeneratedValue(strategy=GenerationType.IDENTITY)
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "ID")
-	private int id;
-	
+	private Long id;
+
 	@NotBlank
 	@Column(name = "STATUS")
 	private String status;
-	
+
 	@NotNull
 	@Column(name = "ORDER_DATE")
 	private LocalDateTime orderDate;
-	
+
 	@NotNull
 	@Column(name = "TB_USER_ID")
-	private int userID;
+	private Long userID;
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -64,12 +63,12 @@ public class Order {
 		this.orderDate = orderDate;
 	}
 
-	public int getUserID() {
+	public Long getUserID() {
 		return userID;
 	}
 
-	public void setUserID(int userID) {
+	public void setUserID(Long userID) {
 		this.userID = userID;
 	}
-	
+
 }
